server/requirements.txt
Flask
flask-cors
PyJWT
Flask-Swagger-UI
pymongo
cloudinary
geopy